\name{HCP}
\alias{HCP}
\alias{HCP.coords}
\docType{data}
\title{ Tree counts in tropical montane forest fragments}
\description{
  \code{HCP} has tree abundance data from 16 forest fragments located in the Highlands of Chiapas, southern Mexico. \code{HCP.coords} 
  contains the geographical UTM coordinates for the 16 forest fragments' centroids.
}
\usage{
data(HCP)
data(HCP.coords)	
}
\format{
  \code{HCP} is a data frame with 231 observations and 19 variables. The three first columns contain family, genus and specific species Latin 
  names. Columns 4 to 19 have tree abundance data for the 16 forest fragments. \code{HCP.coords} is a data frame with two columns and 16 rows.
}
\references{
  Cayuela, L., Golicher, D.J., Rey Benayas, J.M., Gonzalez-Espinosa, M. & Ramirez-Marcial, N. 2006. Fragmentation, disturbance and tree diversity 
  conservation in tropical montane forests. \emph{Journal of Applied Ecology} 43: 1172-1181.
}
\examples{
data(HCP)
data(HCP.coords)
}
\keyword{datasets}
